# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pets API",
    "version": "1.0.0"
  },
  "paths": {
    "/pets": {
      "get": {
        "summary": "Get a list of pets",
        "responses": {
          "200": {
            "description": "A list of pets",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new pet",
        "parameters": [
          {
            "in": "body",
            "name": "pet",
            "description": "Pet to be added",
            "schema": {
              "$ref": "#/definitions/PetCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pet added successfully",
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Referenced user not found"
          }
        }
      }
    },
    "/pets/{id}": {
      "put": {
        "summary": "Update a pet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Pet ID"
          },
          {
            "in": "body",
            "name": "pet",
            "description": "Pet data to update",
            "schema": {
              "$ref": "#/definitions/PetCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet updated successfully"
          }
        }
      },
      "patch": {
        "summary": "Partially update a pet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Pet ID"
          },
          {
            "in": "body",
            "name": "pet",
            "description": "Pet data to partially update",
            "schema": {
              "$ref": "#/definitions/PetCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet partially updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a pet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Pet ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Pet deleted successfully"
          }
        }
      }
    },
    "/users/{id}/pets": {
      "get": {
        "summary": "Get all pets for a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of the user's pets",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          },
          "400": {
            "description": "Invalid user id"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Pet": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string",
          "example": "Bella"
        },
        "type": {
          "type": "string",
          "example": "dog"
        },
        "age": {
          "type": "integer",
          "example": 3
        },
        "userId": {
          "type": "integer",
          "x-nullable": true,
          "example": 1
        },
        "fullname": {
          "type": "string",
          "x-nullable": true,
          "example": "John Doe"
        }
      }
    },
    "PetCreate": {
      "type": "object",
      "required": ["name", "type", "age"],
      "properties": {
        "name": {
          "type": "string",
          "example": "Bella"
        },
        "type": {
          "type": "string",
          "example": "dog"
        },
        "age": {
          "type": "integer",
          "example": 3
        },
        "userId": {
          "type": "integer",
          "x-nullable": true,
          "example": 1
        }
      }
    }
  }
}